requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyjwt
python-dotenv
bcrypt
google-auth
pytest
pytest-asyncio
httpx
websockets
